+++ backend/controllers/programOutcomeController.js
@@ -1,12 +1,15 @@
 const ProgramOutcome = require('../models/ProgramOutcome');
 const Mapping = require('../models/Mapping');
 const Course = require('../models/Course');
+const { applyOwnerScope, claimOwnership } = require('../utils/ownership');
 
 const getProgramOutcomes = async (req, res, next) => {
   try {
     const filter = {};
     if (req.query.courseId) filter.course = req.query.courseId;
-    const outcomes = await ProgramOutcome.find(filter).populate('course', 'name code').sort({ code: 1 });
+    const outcomes = await ProgramOutcome.find(applyOwnerScope(filter, req.user._id))
+      .populate('course', 'name code')
+      .sort({ code: 1 });
     res.json(outcomes);
   } catch (error) {
     next(error);
@@ -20,10 +23,10 @@ const createProgramOutcome = async (req, res, next) => {
       return res.status(400).json({ message: 'code, description and course are required' });
     }
 
-    const courseExists = await Course.findById(course);
+    const courseExists = await Course.findOne(applyOwnerScope({ _id: course }, req.user._id));
     if (!courseExists) return res.status(404).json({ message: 'Parent course not found' });
 
-    const po = await ProgramOutcome.create({ code, description, course });
+    const po = await ProgramOutcome.create({ owner: req.user._id, code, description, course });
     res.status(201).json(po);
   } catch (error) {
     next(error);
@@ -32,12 +35,13 @@ const createProgramOutcome = async (req, res, next) => {
 
 const updateProgramOutcome = async (req, res, next) => {
   try {
-    const po = await ProgramOutcome.findById(req.params.id);
+    const po = await ProgramOutcome.findOne(applyOwnerScope({ _id: req.params.id }, req.user._id));
     if (!po) return res.status(404).json({ message: 'Program outcome not found' });
+    claimOwnership(po, req.user._id);
 
     const { code, description, course } = req.body;
     if (course) {
-      const courseExists = await Course.findById(course);
+      const courseExists = await Course.findOne(applyOwnerScope({ _id: course }, req.user._id));
       if (!courseExists) return res.status(404).json({ message: 'Parent course not found' });
       po.course = course;
     }
@@ -53,10 +57,10 @@ const updateProgramOutcome = async (req, res, next) => {
 
 const deleteProgramOutcome = async (req, res, next) => {
   try {
-    const po = await ProgramOutcome.findById(req.params.id);
+    const po = await ProgramOutcome.findOne(applyOwnerScope({ _id: req.params.id }, req.user._id));
     if (!po) return res.status(404).json({ message: 'Program outcome not found' });
 
-    await Mapping.deleteMany({ programOutcome: po._id });
+    await Mapping.deleteMany(applyOwnerScope({ programOutcome: po._id }, req.user._id));
     await ProgramOutcome.deleteOne({ _id: po._id });
 
     res.json({ message: 'Program outcome and related mappings deleted' });
